Add more plots to end-to-end notebook.
This commit introduces a few new and important plots to the end-to-end
validation notebook. In particular, the notebook now contains a
comparison of real data to simulated data, as well as a single figure
that shows the progression of systematics added to the simulation. The
code for plotting waterfalls has also been removed and now the notebook
uses the fancy plotting tools available in uvtools.