Subtle bugfix in prepare_sim_files script.
There was a bug that caused the script to fall over if the configuration
file had an entry for noise, but did not have a corresponding entry for
setting the random seed. This has been fixed. Note that the bug did not
affect the quality of any of the previous validation simulation
products.